+++ lib/src/modules/reviews/services/fakes/fake_review_service_impl.dart
@@ -10,7 +10,7 @@ class FakeReviewServiceImpl implements IReviewService {
   @override
   Future<void> create({
     required int therapistId,
-    required int rating,
+    required double rating,
     required String comment,
     required DateTime createdAt,
   }) async {
